MERGE: Develop

See merge request s05-webmobile1-sub3/S05P13A608!94

+++ backend/src/main/java/com/calisthenics/homedong/api/service/UserService.java
@@ -111,7 +111,13 @@ public class UserService {
 
     @Transactional
     public void deleteUser() {
-        entryRepositry.deleteByUserId(SecurityUtil.getCurrentEmail().flatMap(userRepository::findOneWithRolesByEmail).orElse(null).getUserId());
+        User user = SecurityUtil.getCurrentEmail().flatMap(userRepository::findOneWithRolesByEmail).orElse(null);
+
+        if(user == null) {
+            throw new UserNotFoundException(SecurityUtil.getCurrentEmail().orElse(""));
+        }
+
+        entryRepositry.deleteByUserId(user.getUserId());
 
         if(userRepository.deleteByEmail(SecurityUtil.getCurrentEmail().orElse("")) == 0) {
             throw new UserNotFoundException(SecurityUtil.getCurrentEmail().orElse(""));
@@ -120,7 +126,19 @@ public class UserService {
 
     @Transactional
     public void deleteUser(final String email) {
-        entryRepositry.deleteByUserId(userRepository.findOneWithRolesByEmail(email).orElse(null).getUserId());
+        User user = SecurityUtil.getCurrentEmail().flatMap(userRepository::findOneWithRolesByEmail).orElse(null);
+
+        if(user == null) {
+            throw new UserNotFoundException(SecurityUtil.getCurrentEmail().orElse(""));
+        }
+
+        User deleteUser = userRepository.findOneWithRolesByEmail(email).orElse(null);
+
+        if(deleteUser == null) {
+            throw new UserNotFoundException(SecurityUtil.getCurrentEmail().orElse(""));
+        }
+
+        entryRepositry.deleteByUserId(deleteUser.getUserId());
 
         if(userRepository.deleteByEmail(email) == 0) {
             throw new UserNotFoundException(SecurityUtil.getCurrentEmail().orElse(""));
